Add options component
