+++ src/components/Pomodoro.tsx
@@ -1,7 +1,7 @@
 import { Grid } from "@mui/material"
 import { Header } from "./Header"
 import { Todo } from "./Todo"
-import { Timer } from "./Timer"
+import { Timer } from "./Timer/Timer"
 
 export const Pomodoro = () => {
     return (
